backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
django-filter
python-dotenv
Pillow
razorpay

.\.venv\Scripts\Activate.ps1